+++ Algorithms-Pathfinder-Project/AdjacencyMatrix.h
@@ -5,7 +5,7 @@ namespace adjMatrix {
 	class AdjacencyMatrix {
 
 	private:
-		int** matrix;
+		double** matrix;
 		int size;
 
 	public:
@@ -13,9 +13,10 @@ namespace adjMatrix {
 		~AdjacencyMatrix();
 		void MakeEmptyGraph(int size);
 		bool IsAdjacent(int u, int v);
-		void AddEdge(int u, int v, int c);
+		void AddEdge(int u, int v, double c);
 		adjList GetAdjList(int u);
 		void RemoveEdge(int u, int v);
+		//void checkValidRoute();
 
 	};
 }
